Split qgscoordinatetransform_p into .h/cpp
